refactor(#337): 클래스 신청 및 패스권 구매 url 병합
