+++ src/app/order/[type]/[id]/_components/Class/Coupon/index.tsx
@@ -1,5 +1,6 @@
+import { CouponSVG } from '@/icons/svg';
 import { getCouponList } from '@/lib/apis/serverApis/couponApis';
-import CouponContainer from './Coupon/CouponContainer';
+import CouponContainer from './CouponContainer';
 import { IcouponsData } from '@/types/coupon';
 
 const Coupon = async ({ id, price }: { id: string; price: number }) => {
@@ -18,7 +19,16 @@ const Coupon = async ({ id, price }: { id: string; price: number }) => {
     return null;
   }
 
-  return <CouponContainer couponList={couponList} price={price} />;
+  return (
+    <section className="mt-4 rounded-md px-4 py-5 shadow-vertical">
+      <h3 className="flex gap-1 text-lg font-semibold">
+        <CouponSVG className="h-6 w-6 fill-sub-color1" />
+        쿠폰 적용
+      </h3>
+
+      <CouponContainer couponList={couponList} price={price} />
+    </section>
+  );
 };
 
 export default Coupon;
