+++ src/app/order/[type]/[id]/_components/Class/Pass/PassSelect.tsx
@@ -1,99 +1,73 @@
 import Select, {
   ActionMeta,
-  FormatOptionLabelMeta,
   MultiValue,
   SingleValue,
   SingleValueProps,
   StylesConfig,
   components,
 } from 'react-select';
-import { SelectCoupon } from '@/types/coupon';
-
-interface CouponSelectProps {
-  options: SelectCoupon[];
+import formatDate from '@/utils/formatDate';
+import { SelectPass } from '@/types/pass';
+/* eslint-disable no-unused-vars */
+interface PassSelectProps {
+  options: SelectPass[];
   onChange: (
-    selectedOptions: MultiValue<SelectCoupon> | SingleValue<SelectCoupon>,
-    actionMeta: ActionMeta<SelectCoupon>,
+    selectedOptions: MultiValue<SelectPass> | SingleValue<SelectPass>,
+    actionMeta: ActionMeta<SelectPass>,
   ) => void;
-  selectValue: SelectCoupon[];
-  type: 'NOMAL' | 'STACKABLE';
+  selectValue: SelectPass[];
 }
-const CouponSelect = ({
-  options,
-  onChange,
-  selectValue,
-  type,
-}: CouponSelectProps) => {
-  const SingleValue = (props: SingleValueProps<SelectCoupon, boolean>) => (
+/* eslint-enable no-unused-vars */
+const PassSelect = ({ options, onChange, selectValue }: PassSelectProps) => {
+  const SingleValue = (props: SingleValueProps<SelectPass, boolean>) => (
     <components.SingleValue {...props}>
-      <div>
-        ({type === 'NOMAL' ? '일반' : '중복'}) {props.data.label}
-      </div>
+      <div>{props.data.label}</div>
     </components.SingleValue>
   );
 
   return (
     <Select
-      instanceId={`${type}coupon-select`}
-      placeholder={`(${type === 'NOMAL' ? '일반' : '중복'}) 적용할 쿠폰 선택`}
-      noOptionsMessage={() =>
-        `적용 가능한 ${type === 'NOMAL' ? '일반' : '중복'}쿠폰이 없습니다`
-      }
+      instanceId="pass-select"
+      placeholder="사용할 패스권 선택"
+      noOptionsMessage={() => `적용 가능한 패스권이 없습니다`}
       formatOptionLabel={formatOptionLabel}
       onChange={onChange}
       options={options}
-      styles={couponSelectStyle}
+      styles={passSelectStyle}
       value={selectValue}
       components={{ SingleValue }}
     />
   );
 };
 
-export default CouponSelect;
+export default PassSelect;
 
-const formatOptionLabel = (
-  { value, label }: SelectCoupon,
-  formatOptionLabelMeta: FormatOptionLabelMeta<SelectCoupon>,
-) => {
-  const {
-    startAt,
-    endAt,
-    percentage,
-    discountPrice,
-    maxDiscountPrice,
-    isStackable,
-  } = value;
+const formatOptionLabel = ({ value, label }: SelectPass) => {
+  const { remainingUses, startAt, endAt } = value;
 
   return (
     <dl className="flex flex-col gap-1 px-3 py-1 text-sm">
       <div className="flex justify-between">
         <dt className="max-w-sm truncate text-black">{label}</dt>
         <dd className="font-semibold text-main-color">
-          {percentage
-            ? percentage + '%'
-            : discountPrice.toLocaleString() + '원'}
+          잔여횟수 {remainingUses}회
         </dd>
       </div>
+
       <div className="flex justify-between">
         <div className="flex gap-2 text-gray-300">
-          <dd>{startAt + '-' + endAt}</dd>
-          {maxDiscountPrice !== 0 && maxDiscountPrice && (
-            <dd>(최대할인 {maxDiscountPrice.toLocaleString()}원)</dd>
+          {startAt && endAt ? (
+            <dd>{`${formatDate(startAt)} - ${formatDate(endAt)}`}</dd>
+          ) : (
+            <dd>미사용</dd>
           )}
         </div>
-        <dd
-          className={` font-semibold ${
-            isStackable ? 'text-main-color' : 'text-gray-300'
-          }`}
-        >
-          {isStackable ? '중복' : '일반'}쿠폰
-        </dd>
       </div>
     </dl>
   );
 };
 
-const couponSelectStyle: StylesConfig<SelectCoupon, true> = {
+const passSelectStyle: StylesConfig<SelectPass, true> = {
   control: (provided) => ({
     ...provided,
     boxShadow: 'none',
