+++ src/app/order/[type]/[id]/_components/Class/Pass/index.tsx
@@ -1,5 +1,6 @@
+import { PassSVG } from '@/icons/svg';
 import { getLecturesPassList } from '@/lib/apis/serverApis/passApis';
-import PassContainer from './Pass/PassContainer';
+import PassContainer from './PassContainer';
 import { SelectPass } from '@/types/pass';
 
 interface PassProps {
@@ -20,7 +21,15 @@ const Pass = async ({ id }: PassProps) => {
     return null;
   }
 
-  return <PassContainer passList={passList} />;
+  return (
+    <section className="mt-4 rounded-md px-4 py-5 shadow-vertical">
+      <h3 className="flex gap-1 text-lg font-semibold">
+        <PassSVG className="h-6 w-6 fill-sub-color1" />
+        패스권
+      </h3>
+      <PassContainer passList={passList} />
+    </section>
+  );
 };
 
 export default Pass;
