+++ src/app/order/[type]/[id]/_components/PaymentType.tsx
