+++ src/app/order/[type]/[id]/loading.tsx
