+++ src/app/order/complete/layout.tsx
@@ -10,9 +10,5 @@ export default function ApplyCompleteLayout({
 }: {
   children: React.ReactNode;
 }) {
-  return (
-    <>
-      {children}
-    </>
-  );
+  return { children };
 }
